backend/requirements.txt
flask
flask-cors
openai-whisper
torch
torchvision
pydub
better_profanity
numpy
werkzeug